Se arreglaron todas las  entidades
Las entidades estaban mal relacionadas

+++ Backend/src/main/java/com/example/Backend/Entity/tiporeporteEntity.java
@@ -12,7 +12,7 @@ public class tiporeporteEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "Id_tipodereporte")
-    private int id_tipodereporte;
+    private Long id_tipodereporte;
 
     @Column(name = "nombre", length = 50, nullable = false)
     private String nombre;
